+++ mobile/src/main/java/be/mygod/vpnhotspot/ui/TetheringScreen.kt
@@ -25,19 +25,10 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import androidx.collection.ScatterSet
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.widthIn
-import androidx.compose.foundation.selection.toggleable
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -45,7 +36,6 @@ import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
-import androidx.compose.material3.VerticalDivider
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
@@ -57,7 +47,6 @@ import androidx.compose.runtime.produceState
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
@@ -67,8 +56,6 @@ import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.semantics.clearAndSetSemantics
-import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.TextLinkStyles
 import androidx.compose.ui.text.buildAnnotatedString
@@ -594,38 +581,11 @@ private fun TetheringRow(
                     onCheckedChange = null,
                 )
             } else {
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Icon(
-                        imageVector = Icons.AutoMirrored.Filled.KeyboardArrowRight,
-                        contentDescription = null,
-                        modifier = Modifier.padding(start = 16.dp, end = 8.dp),
-                        tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                    VerticalDivider(
-                        modifier = Modifier.height(40.dp),
-                        color = MaterialTheme.colorScheme.outlineVariant,
-                    )
-                    Spacer(Modifier.width(12.dp))
-                    Box(
-                        modifier = Modifier
-                            .widthIn(min = 48.dp)
-                            .height(48.dp)
-                            .toggleable(
-                                value = checked,
-                                enabled = switchEnabled,
-                                role = Role.Switch,
-                                onValueChange = { onCheckedChange() },
-                            ),
-                        contentAlignment = Alignment.Center,
-                    ) {
-                        PreferenceSwitch(
-                            checked = checked,
-                            modifier = Modifier.clearAndSetSemantics { },
-                            enabled = switchEnabled,
-                            onCheckedChange = null,
-                        )
-                    }
-                }
+                PreferenceSplitSwitch(
+                    checked = checked,
+                    enabled = switchEnabled,
+                    onCheckedChange = { onCheckedChange() },
+                )
             }
         },
         onClick = onClick,
