+++ mobile/src/main/res/values/strings.xml
@@ -237,11 +237,13 @@
     <string name="wifi_ssid_help">Hotspot SSID in UTF-8.</string>
     <string name="wifi_security">Security</string>
     <string name="wifi_password">Password</string>
+    <string name="wifi_password_error_required">Password is required.</string>
+    <string name="wifi_password_error_length">Password must be 8–63 characters.</string>
     <string name="wifi_password_help">Pre-shared key for WPA2/WPA3 personal security. WPA2-PSK and WPA3-SAE transition passphrases must be 8–63 characters.</string>
     <string name="wifi_hotspot_auto_off">Turn off hotspot automatically</string>
-    <string name="wifi_hotspot_auto_off_help">When no devices are connected for the shutdown timeout.</string>
+    <string name="wifi_hotspot_auto_off_help">When no devices are connected for the shutdown timeout: %s.</string>
     <string name="wifi_hotspot_timeout">Shutdown timeout</string>
-    <string name="wifi_hotspot_timeout_help">Time to wait with no connected devices before auto shutdown.&lt;br&gt;&lt;br&gt;Default: %s.</string>
+    <string name="wifi_hotspot_timeout_help">Configure the time to wait with no connected devices before auto shutdown.&lt;br&gt;&lt;br&gt;Default: %s.</string>
     <string name="wifi_hotspot_timeout_default">Default (%s)</string>
     <string name="wifi_hotspot_ap_band_title">AP Band</string>
     <string name="wifi_hotspot_ap_channel_band_title">AP channel/band</string>
@@ -249,7 +251,7 @@
     <string name="wifi_hotspot_concurrent_ap_channel_band_title">Concurrent AP channel/band</string>
     <string name="wifi_hotspot_concurrent_ap_channel_band_help">Adds a second AP instance when bridged AP concurrency is supported. Channel &lt;tt&gt;0&lt;/tt&gt; lets Android auto-select. Unsupported band/channel combinations can fail when the hotspot starts.</string>
     <string name="wifi_band_optimization">Enable band optimization</string>
-    <string name="wifi_band_optimization_help">Lets Android optimize band settings, such as appending the 2.4 GHz band for coexistence or country-code restrictions. AOSP enables this by default.</string>
+    <string name="wifi_band_optimization_help">Lets Android optimize band settings, such as appending the 2.4 GHz band for coexistence or country-code restrictions.</string>
     <string name="wifi_ap_choose_disabled">Disabled</string>
     <string name="wifi_ap_choose_G">%s GHz Band</string>
     <string name="wifi_hotspot_acs_channel_2g">Allowed 2.4 GHz ACS channels</string>
@@ -283,9 +285,9 @@
     <string name="wifi_mac_randomization_non_persistent">Non-persistent</string>
     <string name="wifi_mac_randomization_non_persistent_help">Generate a randomized AP MAC address each time the AP is re-enabled.</string>
     <string name="wifi_bridged_mode_opportunistic_shutdown">Bridged mode opportunistic shutdown</string>
-    <string name="wifi_bridged_mode_opportunistic_shutdown_help">In bridged AP mode, Android can shut down an idle AP instance after the bridged timeout. If both instances are idle, Android shuts down the higher-frequency instance.</string>
+    <string name="wifi_bridged_mode_opportunistic_shutdown_help">In bridged AP mode, Android can shut down an idle AP instance after the bridged timeout: %s.</string>
     <string name="wifi_hotspot_timeout_bridged">Inactive timeout for a bridged instance</string>
-    <string name="wifi_hotspot_timeout_bridged_help">Time an idle bridged AP instance waits before opportunistic shutdown. The overall hotspot timeout still applies separately.&lt;br&gt;&lt;br&gt;Default: %s.</string>
+    <string name="wifi_hotspot_timeout_bridged_help">Configure the time an idle bridged AP instance waits before opportunistic shutdown. If both instances are idle, Android shuts down the higher-frequency instance. The overall hotspot timeout still applies separately.&lt;br&gt;&lt;br&gt;Default: %s.</string>
     <string name="wifi_ieee_80211ax">Enable Wi\u2011Fi 6</string>
     <string name="wifi_ieee_80211ax_help">Enable 802.11ax for Soft AP when the device supports it.</string>
     <string name="wifi_ieee_80211be">Enable Wi\u2011Fi 7</string>
